Moves "Select" (ok) buttons to the right and adds an icon
When selecting an input or output model, the "Select" buttons would appear at the very left side of the window, which is a bit irritating,
This change adds a horizontal layout with a stretcher to the left, such that the buttons appear on the far right side.
Additionally, the buttons also get a green checkmark icon now, so that they are more visually distintitve.