Use the conversion name as the output file name
So far, the output file name was automatically generated from the input file name. This prevents people from doing multiple conversions of the same mesh with different parameters.

Since each conversion has a user selected name already, anyway, it makes sense to use that as the output file name. If no name is provided, the system already sets up a default name from the input meshes' name, so in that case it works just as before.

Additionally, selecting the input mesh would not update the list view on the left, because that information was not communicated to the UI. Sending the proper signal makes the UI update correctly.